Implementando gráfico na pagina de visualizar clientes

+++ src/app/views/view-cars/view-cars.component.ts
@@ -2,12 +2,13 @@ import { Component, OnInit } from '@angular/core';
 import { Observable } from 'rxjs';
 import { Car } from 'src/app/models/car.model';
 import CarService from 'src/app/services/car.service';
+import MessageService from '../../services/message.service';
 
 @Component({
   selector: 'app-view-cars',
   templateUrl: './view-cars.component.html',
   styleUrls: ['./view-cars.component.scss'],
-  providers:[CarService]
+  providers: [CarService, MessageService]
 })
 export class ViewCarsComponent implements OnInit {
   public cars !: Observable<Car[]>
@@ -15,16 +16,21 @@ export class ViewCarsComponent implements OnInit {
   public displayedColumns: string[] = ['client','plate' ,'colorName' ,'model', 'brand' ,'price', 'actions'];
 
   constructor(
-    private readonly carService : CarService
+    private readonly carService: CarService,
+    private readonly messageService: MessageService
     ) { }
 
   ngOnInit(): void {
     this.loadCars()
   }
 
-  public confirmDeleteCar(id : string){
-    this.carService.deleteCar(id)
-    .subscribe(() => this.loadCars())
+  public confirmDeleteCar(id: string) {
+    this.messageService.showConfirmDialog("Deseja mesmo excluir as informações desse carro?", "")
+      .then(result => {
+        if (result.isConfirmed) this.carService.deleteCar(id)
+          .subscribe(() => this.loadCars())
+      })
+
 
   }
